# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QueueLess API",
    "description": "Predict waiting time at government offices + remote virtual tokens",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/citizen/offices/{office_id}/services": {
      "get": {
        "tags": [
          "Citizen"
        ],
        "summary": "List Office Services",
        "operationId": "list_office_services_v1_citizen_offices__office_id__services_get",
        "parameters": [
          {
            "name": "office_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Office Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceOut"
                  },
                  "title": "Response List Office Services V1 Citizen Offices  Office Id  Services Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/citizen/tokens": {
      "post": {
        "tags": [
          "Citizen"
        ],
        "summary": "Create Token",
        "operationId": "create_token_v1_citizen_tokens_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenBookIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/citizen/tokens/me/active": {
      "get": {
        "tags": [
          "Citizen"
        ],
        "summary": "Get My Active Token",
        "operationId": "get_my_active_token_v1_citizen_tokens_me_active_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TokenOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get My Active Token V1 Citizen Tokens Me Active Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/citizen/tokens/{token_id}": {
      "get": {
        "tags": [
          "Citizen"
        ],
        "summary": "Get Token Details",
        "operationId": "get_token_details_v1_citizen_tokens__token_id__get",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/citizen/tokens/{token_id}/cancel": {
      "post": {
        "tags": [
          "Citizen"
        ],
        "summary": "Citizen Cancel Token",
        "operationId": "citizen_cancel_token_v1_citizen_tokens__token_id__cancel_post",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/citizen/tokens/{token_id}/check-in": {
      "post": {
        "tags": [
          "Citizen"
        ],
        "summary": "Citizen Check In",
        "operationId": "citizen_check_in_v1_citizen_tokens__token_id__check_in_post",
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/citizen/devices": {
      "post": {
        "tags": [
          "Citizen"
        ],
        "summary": "Register Device",
        "operationId": "register_device_v1_citizen_devices_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/citizen/me": {
      "patch": {
        "tags": [
          "Citizen"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_v1_citizen_me_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/devices": {
      "post": {
        "tags": [
          "Core V1"
        ],
        "summary": "Register Device",
        "description": "C7: Register FCM device token and language for push notifications.\nUpserts into devices table.",
        "operationId": "register_device_v1_devices_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me": {
      "patch": {
        "tags": [
          "Core V1"
        ],
        "summary": "Update Profile",
        "description": "C11: Update profile / language. Server notification templates follow it.",
        "operationId": "update_profile_v1_me_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckInIn": {
        "properties": {
          "qr_payload": {
            "type": "string",
            "title": "Qr Payload"
          }
        },
        "type": "object",
        "required": [
          "qr_payload"
        ],
        "title": "CheckInIn"
      },
...
      "DeviceRegisterIn": {
        "properties": {
          "fcm_token": {
            "type": "string",
            "title": "Fcm Token"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "android"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "fcm_token"
        ],
        "title": "DeviceRegisterIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileUpdateIn": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "ProfileUpdateIn"
      },
...
      "ServiceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "office_id": {
            "type": "string",
            "title": "Office Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "names": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Names"
          },
          "prior_avg_minutes": {
            "type": "number",
            "title": "Prior Avg Minutes"
          },
          "required_docs": {
            "items": {},
            "type": "array",
            "title": "Required Docs"
          },
          "priority_allowed": {
            "type": "boolean",
            "title": "Priority Allowed"
          },
          "requires_physical_visit": {
            "type": "boolean",
            "title": "Requires Physical Visit"
          },
          "online_alternative_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Online Alternative Url"
          },
          "location_hint": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Hint"
          },
          "indicative_wait_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indicative Wait Minutes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "office_id",
          "code",
          "names",
          "prior_avg_minutes",
          "required_docs",
          "priority_allowed",
          "requires_physical_visit"
        ],
        "title": "ServiceOut"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Operation completed successfully"
          }
        },
        "type": "object",
        "title": "SuccessResponse"
      },
      "TokenBookIn": {
        "properties": {
          "office_id": {
            "type": "string",
            "title": "Office Id"
          },
          "service_id": {
            "type": "string",
            "title": "Service Id"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "NORMAL"
          },
          "priority_doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Doc Type"
          },
          "beneficiary_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Name"
          },
          "travel_minutes": {
            "type": "integer",
            "title": "Travel Minutes",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "office_id",
          "service_id",
          "phone"
        ],
        "title": "TokenBookIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}